+++ src/front/default/app.cljs
@@ -1,12 +1,17 @@
 (ns default.app
   (:require [reagent.dom :as rdom]
-            [goog.string :as gstr]))
+            [goog.string :as gstr]
+            [default.views :as views]
+            [default.events :as events]
+            [default.subs :as subs]
+            ))
 
 (defn app-page []
   [:div#someId.someClass
    [:span.nestedClass
     "hello," (gstr/unescapeEntities "&nbsp;") "World !" ;; use HTML entities
     ]
+   [views/say-hi-widget]
    [:div
     [:img.clojureLogo {:src "image/clojure-logo.png"}]]])