+++ src/front/default/subs.cljs
@@ -1,0 +1,8 @@
+(ns default.subs
+  (:require [re-frame.core :as rf]))
+
+;; layer 2
+
+(rf/reg-sub  :username
+             (fn [db _]
+               (:username db)))