+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning].
   - add dependency to [Re-frame](https://day8.github.io/re-frame/)
   - add dependency to [re-frame-10x](https://github.com/day8/re-frame-10x)
   - add dependency to [cljs-devtools](https://github.com/binaryage/cljs-devtools)
+  - add dependency to [Bulma](https://bulma.io) (pure CSS Framework)
 
 ## [0.0.2]
 - ready to use dev environment and tooling (see [README](README.md) for details)